Make changes to stdlib to make it more Dart-like

+++ compiler/src/test/kotlin/compile/Object.kt
@@ -52,7 +52,7 @@ class Object : BaseTest {
         kotlin(
             """
             object Test {
-                fun compute(x: Int, y: Float): Int {
+                fun compute(x: Int, y: Double): Int {
                     return 343
                 }
             }
